initial impl for mmap, needs cleanup, bug fixes
